Add files for Matt's example
Mathematica notebook Generate rotations.nb generates a C-like file which needs to be converted to true C form with sed in the script CFormToC.sh. The resulting C code (in quaternion_formulae.c) is included in rotation_example.c, along with twister.c (which uses the file ISEED to generate random numbers).

rotation_example.c generates a series of random rotations which rotate the point (1,0,0) about the origin.